Refactor/Ajuste na tela de configurações
- Mudamos as actions de privacidade/disponivel para trabalho do profiles_controller
para o settings_controller

- Também foi alterada a localização dos botões nas views
