+++ app/controllers/settings_controller.rb
@@ -15,6 +15,28 @@ class SettingsController < ApplicationController
     redirect_to root_path, notice: t('delete_account')
   end
 
+  def change_privacy
+    @profile = current_user.profile
+    if @profile.public_profile?
+      @profile.private_profile!
+    else
+      @profile.public_profile!
+    end
+    redirect_to profile_settings_path(@profile), notice: t('.success')
+  end
+
+  def work_unavailable
+    @profile = current_user.profile
+    @profile.unavailable!
+    redirect_to profile_settings_path(@profile), notice: t('.success')
+  end
+
+  def open_to_work
+    @profile = current_user.profile
+    @profile.open_to_work!
+    redirect_to profile_settings_path(@profile), notice: t('.success')
+  end
+
   private
 
   def redirect_unauthorized_access
