+++ app/views/settings/index.html.erb
@@ -1,20 +1,44 @@
-<h2>Configurações</h2>
-<div>
-	<%= form_with(url: deactivate_profile_path, method: :patch) do |form| %>
-		<p><strong class="text-danger">Atenção:</strong> Todos os dados relacionados ao seu perfil serão ARQUIVADOS. Você pode reativar sua conta fazendo log-in novamente.</p>
-		<%= form.submit 'Desativar Perfil',
-			class: "btn btn-danger",
-			data: { turbo_confirm: 'Seu perfil será DESATIVADO, deseja prosseguir?' }
-		%>
-	<% end %>
-</div>
+<h2 class="text-center">Configurações</h2>
 
-<div>
-	<%= form_with(url: delete_account_path, method: :delete) do |form| %>
-		<p><strong class="text-danger">Atenção:</strong> Todos os dados relacionados a conta serão EXCLUÍDOS e não poderão ser recuperados.</p>
-		<%= form.submit 'Excluir Conta',
-			class: "btn btn-danger",
-			data: { turbo_confirm: 'Sua conta será EXCLUÍDA, deseja prosseguir?' }
-		%>
-	<% end %>
-</div>
+<section class="text-center mt-4">	
+	<h3>Privacidade</h3>
+	<div class="mt-3">
+		<% if @profile.public_profile? %>
+			<p class="text-success fs-4"><%= t Profile.human_attribute_name("privacy.#{@profile.privacy}") %></p>
+		<% else %>
+			<p class="text-secondary fs-4"><%= t Profile.human_attribute_name("privacy.#{@profile.privacy}") %></p>
+		<% end %>
+		<%= button_to 'Alterar Privacidade', change_privacy_path, method: :patch, class: 'btn btn-outline-primary btn-sm' %>
+	</div>
+	<div class="mt-2">
+		<% if @profile.open_to_work? %>
+			<p class="text-success fs-4"><%= t Profile.human_attribute_name("work_status.#{@profile.work_status}") %></p>
+			<%= button_to 'Alterar Disponibilidade', work_unavailable_path, method: :patch, class: 'btn btn-outline-primary btn-sm' if current_user.profile == @profile %>
+		<% else %>
+			<p class="text-danger fs-4"><%= t Profile.human_attribute_name("work_status.#{@profile.work_status}") %></p>
+			<%= button_to 'Alterar Disponibilidade', open_to_work_path, method: :patch, class: 'btn btn-outline-primary btn-sm' if current_user.profile == @profile %>
+		<% end %>
+	</div>
+</section>
+<br>
+<h3 class="text-center">Remoção de conta</h3>
+<p class="text-center"><strong class="text-danger">Atenção:</strong> Ao desativar o perfil todos os dados relacionados serão <strong>ARQUIVADOS</strong>. Você pode reativar sua conta fazendo login novamente.</p>
+<section class="text-center d-flex justify-content-around w-25 mx-auto">
+	<div>
+		<%= form_with(url: deactivate_profile_path, method: :patch) do |form| %>
+			<%= form.submit 'Desativar Perfil',
+				class: "btn btn-danger",
+				data: { turbo_confirm: 'Seu perfil será DESATIVADO, deseja prosseguir?' }
+			%>
+		<% end %>
+	</div>
+
+	<div>
+		<%= form_with(url: delete_account_path, method: :delete) do |form| %>
+			<%= form.submit 'Excluir Conta',
+				class: "btn btn-danger",
+				data: { turbo_confirm: 'Sua conta será EXCLUÍDA, deseja prosseguir?' }
+			%>
+		<% end %>
+	</div>
+</section>
