+++ config/locales/profile.pt-BR.yml
@@ -27,12 +27,6 @@ pt-BR:
       results:
         one: 'resultado'
         other: 'resultados'
-    work_unavailable:
-      success: Alteração salva com sucesso
-    open_to_work:
-      success: Alteração salva com sucesso
-    change_privacy:
-      success: Privacidade alterada com sucesso
     show:
       private: O perfil que você acessou é privado
   profile:
