+++ config/routes.rb
@@ -24,7 +24,7 @@ Rails.application.routes.draw do
   resources :users, only: [] do
     resources :posts, shallow: true, only: %i[show edit update]
     resources :profiles, shallow: true, only: %i[edit show update] do
-      resources 'settings', only: %i[index]
+      resources :settings, only: %i[index]
       patch :remove_photo, on: :member
       resources :connections, only: %i[create index] do
         patch 'unfollow', 'follow_again'
@@ -34,10 +34,7 @@ Rails.application.routes.draw do
   end
 
   delete 'delete_account', controller: :settings
-  patch 'deactivate_profile', controller: :settings
-  patch 'work_unavailable', controller: :profiles
-  patch 'open_to_work', controller: :profiles
-  patch 'change_privacy', controller: :profiles
+  patch 'deactivate_profile', 'work_unavailable', 'open_to_work', 'change_privacy', controller: :settings
 
   resources :likes, only: %i[create destroy]
   resources :job_categories, only: %i[index create]
